requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
altair==4.1.0
appdirs==1.4.4
appnope==0.1.2
argh==0.26.2
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
base58==2.1.0
bleach==3.3.0
blinker==1.4
boto3==1.17.109
botocore==1.20.109
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cryptography==3.4.7
cycler==0.10.0
cytoolz==0.11.0
dask==2021.7.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
flatbuffers==2.0
fsspec==2021.6.1
gast==0.5.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.32.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.38.1
h5py==2.10.0
idna==2.10
imagecodecs==2021.6.8
imageio==2.9.0
importlib-metadata==4.6.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
locket==0.2.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.3
notebook==6.4.0
numpy==1.21.0
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==20.2.4
pooch==1.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.16.0
ptyprocess==0.7.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pydeck==0.6.2
Pygments==2.9.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
pyu2f==0.1.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
Send2Trash==1.7.1
setuptools==49.6.0.post20210108
six==1.16.0
smmap==3.0.5
statsmodels==0.12.2
streamlit==0.81.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons-macos==0.1a3
tensorflow-estimator==2.5.0
tensorflow-macos==0.1a3
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.7.2
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typeguard==2.12.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0
